# frontend: Speed up repository listing by increasing batch size