--- a/cmd/frontend/graphqlbackend/repositories.go
+++ b/cmd/frontend/graphqlbackend/repositories.go
@@ -3,6 +3,7 @@ package graphqlbackend
 import (
 	"context"
 	"regexp"
+	"sort"
 	"strings"
 	"sync"
 	"time"
@@ -98,7 +99,14 @@ func (r *repositoryConnectionResolver) compute(ctx context.Context) ([]*types.Re
 		if opt2.LimitOffset != nil {
 			tmp := *opt2.LimitOffset
 			opt2.LimitOffset = &tmp
-			opt2.Limit++ // so we can detect if there is a next page
+			// We purposefully load more repos into memory than requested in
+			// order to save roundtrips to gitserver in case we need to do
+			// filtering by clone status.
+			// The trade-off here is memory/cpu vs. network roundtrips to
+			// database/gitserver and we choose smaller latency over smaller
+			// memory footprint
+			// At the end of this method we return the requested number of repos
+			opt2.Limit += 500
 		}
 
 		var indexed map[api.RepoName]bool
@@ -165,13 +173,22 @@ func (r *repositoryConnectionResolver) compute(ctx context.Context) ([]*types.Re
 			}
 
 			r.repos = append(r.repos, repos...)
+			// TODO(@mrnugget): This needs to take the sort-column into account
+			sort.Slice(r.repos, func(i, j int) bool { return r.repos[i].Name < r.repos[j].Name })
+
 			if opt2.LimitOffset == nil {
 				break
 			} else {
-				if len(r.repos) >= opt2.Limit {
+				if len(r.repos) >= r.opt.Limit {
+					// Cut off the repos we additionally loaded to save
+					// roundtrips to `gitserver` and only return the number
+					// that was requested.
+					// But, when possible, we add one more so we can detect if
+					// there is a "next page" that could be loaded
+					r.repos = r.repos[:r.opt.Limit+1]
 					break
 				}
-				if reposFromDB < opt2.Limit {
+				if reposFromDB < r.opt.Limit {
 					break
 				}
 				opt2.Offset += opt2.Limit