@tsenart 2019-06-01
@keegancsmith, @mrnugget: What if we loaded all cloned repos from git-server first with a single request, and then would filter in-memory? How slow does that loading of all cloned repo names from git-server need to be?