@mrnugget 2019-06-03
> @mrnugget noticed an easy fix for the sorting, so pushed it up.

Thank you! That's the most elegant solution — fixing the problem at the root :)

> @keegancsmith, @mrnugget: What if we loaded all cloned repos from git-server first with a single request, and then would filter in-memory?

As things currently stand, I think that would only get us halfway there, since `gitserver` only has an [endpoint](https://github.com/sourcegraph/sourcegraph/blob/f54077a3f02f47fd81dc62285d2aab7c43dcadb4/cmd/gitserver/server/list.go#L10-L87) to return a list of names of cloned repositories. What's missing is whether the repo is `CloneInProgress`.

So if we use the endpoint as it is, we can use it as a first optimization, but if the user asks for the `CloneInProgress` filter, we would need to ask `gitserver`. At the moment I personally don't think introducing that codepath is worth it.

> How slow does that loading of all cloned repo names from git-server need to be?

I have no idea about the performance of `gitserver` and since I basically pulled the batch size of `500` out of magic hat, I was wondering: what's the biggest batch size we should send to `gitserver`?

Consider that the `gitserver` client [shards](https://github.com/sourcegraph/sourcegraph/blob/f54077a3f02f47fd81dc62285d2aab7c43dcadb4/pkg/gitserver/client.go#L561-L587) but that on each `gitserver` each repository is handled sequentially and to construct the complete `RepoInfo` response, a few filesystem operations are necessary: https://github.com/sourcegraph/sourcegraph/blob/f54077a3f02f47fd81dc62285d2aab7c43dcadb4/cmd/gitserver/server/repo_info.go#L17-L57

I'd say: filesystems are fast as hell, PostgreSQL is too, we could up the batch size to `1000`. What's your opinion on this, @keegancsmith @slimsag @tsenart?