@mrnugget 2019-06-03
(Merging this and I'll follow up with possible further optimizations in other PRs)