@tsenart 2019-06-03
> (Merging this and I'll follow up with possible further optimizations in other PRs)

👍 

> I'd say: filesystems are fast as hell, PostgreSQL is too, we could up the batch size to 1000. What's your opinion on this, @keegancsmith @slimsag @tsenart?

I'd say we benchmark!